pkg/fms: Migrate config items to pkg/config
