+++ internal/cmdlets/fms_run.go
@@ -13,8 +13,9 @@ import (
 
 	"github.com/spf13/cobra"
 
+	"github.com/gizmo-platform/gizmo/pkg/config"
 	"github.com/gizmo-platform/gizmo/pkg/fms"
-	"github.com/gizmo-platform/gizmo/pkg/routeros/config"
+	rconfig "github.com/gizmo-platform/gizmo/pkg/routeros/config"
 	"github.com/gizmo-platform/gizmo/pkg/tlm/net"
 )
 
@@ -41,17 +42,17 @@ func fmsRunCmdRun(c *cobra.Command, args []string) {
 
 	wg := new(sync.WaitGroup)
 
-	fmsConf, err := fms.NewConfig(appLogger)
+	fmsConf, err := config.NewFMSConfig(appLogger)
 	if err != nil {
 		appLogger.Error("Could not load fms.json, have you run the wizard yet?", "error", err)
 		return
 	}
 
 	routerAddr := "100.64.0.1"
-	controller := config.New(
-		config.WithFMS(fmsConf),
-		config.WithLogger(appLogger),
-		config.WithRouter(routerAddr),
+	controller := rconfig.New(
+		rconfig.WithFMS(fmsConf),
+		rconfig.WithLogger(appLogger),
+		rconfig.WithRouter(routerAddr),
 	)
 	appLogger.Debug("Controller Init")
