+++ pkg/fms/fms.go
@@ -167,7 +167,7 @@ func (f *FMS) filterSplit(in, param *pongo2.Value) (*pongo2.Value, *pongo2.Error
 }
 
 func (f *FMS) filterTeamName(in, param *pongo2.Value) (*pongo2.Value, *pongo2.Error) {
-	t, ok := in.Interface().(*Team)
+	t, ok := in.Interface().(*config.Team)
 	if !ok {
 		f.l.Error("Something that wasn't a team got passed to the teamName filter", "in", in.Interface())
 		return pongo2.AsValue(""), &pongo2.Error{Sender: "filter:teamName"}
